+++ src/App.tsx
@@ -2,7 +2,11 @@ import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import { ThemeProvider } from "@/components/theme-provider";
 import { Toaster } from "@/components/ui/sonner";
 import Layout from "@/components/layout";
-import RepoView from "@/components/repo-view";
+import RepoView, {
+  LotteryFactorRoute,
+  ContributionsRoute,
+  DistributionRoute,
+} from "@/components/repo-view";
 import Home from "@/components/home";
 
 function App() {
@@ -13,12 +17,9 @@ function App() {
           <Route path="/" element={<Layout />}>
             <Route index element={<Home />} />
             <Route path="/:owner/:repo" element={<RepoView />}>
-              <Route path="" element={<RepoView.LotteryFactor />} />
-              <Route
-                path="contributions"
-                element={<RepoView.Contributions />}
-              />
-              <Route path="distribution" element={<RepoView.Distribution />} />
+              <Route path="" element={<LotteryFactorRoute />} />
+              <Route path="contributions" element={<ContributionsRoute />} />
+              <Route path="distribution" element={<DistributionRoute />} />
             </Route>
           </Route>
         </Routes>
